[Refactor] 리뷰 반영
