+++ ios-contact-manager/ios-contact-manager/Controller/ContactTableViewCell.swift
@@ -1,21 +1,19 @@
-//
-//  ContactTableViewCell.swift
-//  ios-contact-manager
-//
-//  Created by Swain Yun on 10/5/23.
-//
-
 import UIKit
 
-class ContactTableViewCell: UITableViewCell {
+final class ContactTableViewCell: UITableViewCell {
 
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
     }
 
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
     }
-
+    
+    func configureCell(item contact: Contact) {
+        var content = self.defaultContentConfiguration()
+        content.text = "\(contact.name)(\(contact.age))"
+        content.secondaryText = "\(contact.phoneNumber)"
+        self.contentConfiguration = content
+    }
 }
