+++ ios-contact-manager/ios-contact-manager/Models/Contact.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct Contact {
+    let name: String
+    let age: Int
+    let phoneNumber: Int
+}
